sistema_bi_salud_publica/requirements.txt
pandas>=2.0.0
pyodbc>=5.0.0
streamlit>=1.28.0
plotly>=5.18.0
python-dotenv>=1.0.0
